revert removing IMethodAspectHandler. This is needed for async method parsing because aspect injector does not support it by default
