+++ libraries/src/AWS.Lambda.Powertools.Tracing/TracingAttribute.cs
@@ -13,8 +13,6 @@
  * permissions and limitations under the License.
  */
 
-using System;
-using AspectInjector.Broker;
 using AWS.Lambda.Powertools.Common;
 using AWS.Lambda.Powertools.Tracing.Internal;
 
@@ -107,8 +105,7 @@ namespace AWS.Lambda.Powertools.Tracing;
 ///         }
 ///     </code>
 /// </example>
-[Injection(typeof(TracingAspect))]
-public class TracingAttribute : Attribute
+public class TracingAttribute : MethodAspectAttribute
 {
     /// <summary>
     ///     Set custom segment name for the operation.
@@ -131,4 +128,20 @@ public class TracingAttribute : Attribute
     /// </summary>
     /// <value>The capture mode.</value>
     public TracingCaptureMode CaptureMode { get; set; } = TracingCaptureMode.EnvironmentVariable;
+
+    /// <summary>
+    ///     Creates the handler.
+    /// </summary>
+    /// <returns>IMethodAspectHandler.</returns>
+    protected override IMethodAspectHandler CreateHandler()
+    {
+        return new TracingAspectHandler
+        (
+            SegmentName,
+            Namespace,
+            CaptureMode,
+            PowertoolsConfigurations.Instance,
+            XRayRecorder.Instance
+        );
+    }
 }
